+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * Default vmaf n_threads to the number of logical CPUs.
 * Add `--temp-dir` argument to specify storage of sample data. 
   May also be set with env var `AB_AV1_TEMP_DIR`.
+* Add `--sample-every DURATION` argument, default "12m".
+* Remove 3 sample default, this is now calculated using `--sample-every` 12m default.
 * Create samples concurrently while encoding to reduce io lags waiting to encode.
 * _crf-search_ re-use samples for crf analysis.
 * Linux: _vmaf_ use fifo to convert both reference & distorted to yuv which fixes vmaf accuracy in some cases.